changed colors, component modularization

+++ src/app/shared/components/contact-card/contact-card.component.ts
@@ -1,6 +1,7 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { ContactsService } from 'src/app/contacts/service/contacts.service';
 import { Contact } from 'src/app/models/contact.interface';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-contact-card',
@@ -9,11 +10,14 @@ import { Contact } from 'src/app/models/contact.interface';
 })
 export class ContactCardComponent implements OnInit {
   @Input() contact: Contact = {'id':-1,'first_name':'','last_name':'','email':'','telephone':''};
+  @Output() messageEvent = new EventEmitter<string>();
   contacts:Contact[] = [];
+  message: string = "Hola Mundo!";
   
   constructor(private _contact_service: ContactsService) { }
 
   ngOnInit(): void {
+    
   }
 
   get_contacts():Contact[] {
@@ -22,12 +26,42 @@ export class ContactCardComponent implements OnInit {
     
   }
 
-
   delete_contact(id:number):void {
     // console.log(id);
-    this._contact_service.delete_contact(id);
-    this.get_contacts();
+    Swal.fire({
+      title: 'Are you sure want to remove?',
+      text: 'You will not be able to recover this contact!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        this._contact_service.delete_contact(id);
+        this.messageEvent.emit(this.message)
+        Swal.fire(
+          'Deleted!',
+          'Your contact has been deleted.',
+          'success'
+        )
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your contact is safe :)',
+          'error'
+        )
+      }
+    })
+    
     
   }
+
+
+  // delete_contact(id:number):void {
+  //   // console.log(id);
+  //   this._contact_service.delete_contact(id);
+  //   this.messageEvent.emit(this.message)
+    
+  // }
 
 }
